Fix overflow of 32-bit signed integers when multiplying two 32-bit integers: cast to 64-bit integers first, otherwise matrices with number of elements > 2^31 (e.g. a 48,000x48,000 custom kernel matrix) will not work.

+++ src/shogun/lib/SGNDArray.h
@@ -46,7 +46,7 @@ template<class T> class SGNDArray : public SGReferencedData
 		T* get_matrix(index_t matIdx) const
 		{
 			ASSERT(array && dims && num_dims > 2 && dims[2] > matIdx)
-			return &array[matIdx*dims[0]*dims[1]];
+			return &array[int64_t(matIdx)*int64_t(dims[0])*dims[1]];
 		}
 
 		/** transposes a matrix formed by the two first dimensions
